refactor: extract shared OAuth CLI helpers into oauth/shared.ts (#21375)
