+++ assistant/src/cli/commands/oauth/platform.ts
@@ -6,26 +6,19 @@ import {
   ServicesSchema,
 } from "../../../config/schemas/services.js";
 import { getProvider } from "../../../oauth/oauth-store.js";
-import { VellumPlatformClient } from "../../../platform/client.js";
 import { openInBrowser } from "../../../util/browser.js";
 import { getCliLogger } from "../../logger.js";
 import { shouldOutputJson, writeOutput } from "../../output.js";
+import {
+  fetchActiveConnections,
+  requirePlatformClient,
+  toBareProvider,
+} from "./shared.js";
 
 const log = getCliLogger("cli");
 
 // ---------------------------------------------------------------------------
-// Shared types
-// ---------------------------------------------------------------------------
-
-interface PlatformConnectionEntry {
-  id: string;
-  account_label?: string;
-  scopes_granted?: string[];
-  status?: string;
-}
-
-// ---------------------------------------------------------------------------
-// Shared helpers
+// Platform-specific helpers
 // ---------------------------------------------------------------------------
 
 /**
@@ -36,17 +29,6 @@ function toProviderKey(provider: string): string {
   return provider.startsWith("integration:")
     ? provider
     : `integration:${provider}`;
-}
-
-/**
- * Extract the bare provider slug (e.g. "google") from either a raw CLI
- * argument or a canonical provider key (e.g. "integration:google").
- * Platform API paths expect the bare slug, not the internal key.
- */
-function toBareProvider(provider: string): string {
-  return provider.startsWith("integration:")
-    ? provider.slice("integration:".length)
-    : provider;
 }
 
 /**
@@ -99,60 +81,6 @@ function requireManagedProvider(provider: string, cmd: Command): string | null {
   }
 
   return managedKey;
-}
-
-/**
- * Create an authenticated platform client, or write an error and return null.
- */
-async function requirePlatformClient(
-  cmd: Command,
-): Promise<VellumPlatformClient | null> {
-  const client = await VellumPlatformClient.create();
-  if (!client || !client.platformAssistantId) {
-    writeOutput(cmd, {
-      ok: false,
-      error:
-        "Platform prerequisites not met (not logged in or missing assistant ID)",
-    });
-    process.exitCode = 1;
-    return null;
-  }
-  return client;
-}
-
-/**
- * Fetch active platform connections for a provider. Returns the parsed entries
- * or writes an error and returns null.
- */
-async function fetchActiveConnections(
-  client: VellumPlatformClient,
-  provider: string,
-  cmd: Command,
-): Promise<PlatformConnectionEntry[] | null> {
-  const params = new URLSearchParams();
-  params.set("provider", toBareProvider(provider));
-  params.set("status", "ACTIVE");
-
-  const path = `/v1/assistants/${encodeURIComponent(client.platformAssistantId)}/oauth/connections/?${params.toString()}`;
-  const response = await client.fetch(path);
-
-  if (!response.ok) {
-    writeOutput(cmd, {
-      ok: false,
-      error: `Platform returned HTTP ${response.status}`,
-    });
-    process.exitCode = 1;
-    return null;
-  }
-
-  const body = (await response.json()) as unknown;
-
-  // The platform returns either a flat array or a {results: [...]} wrapper.
-  return (
-    Array.isArray(body)
-      ? body
-      : ((body as Record<string, unknown>).results ?? [])
-  ) as PlatformConnectionEntry[];
 }
 
 // ---------------------------------------------------------------------------
